reviewers comments resolved

+++ app/src/main/java/za/co/woolworths/financial/services/android/firebase/model/ConfigData.kt
@@ -2,6 +2,6 @@ package za.co.woolworths.financial.services.android.firebase.model
 
 data class ConfigData(
     val activeConfiguration: ActiveConfiguration,
-    val expiryTime: Int,
+    val expiryTime: Long,
     val inactiveConfiguration: InactiveConfiguration
 )
